# ast+topdown: move all-ground-refs check into index build