Previously, we'd end up iterating and checking all rules BEFORE index lookup. Now, the groundness check is moved into index creation time, where it only happens once, and is cheap to retrieve.